Expand packet IDs to 64-bit
The LoRa packet layout now uses a 64-bit collar_id instead of a 16-bit value, increasing the total packet size from 17 to 23 bytes. This updates the firmware packet structs, base station and collar ID formatting, and the project documentation so longer hexadecimal identifiers are transmitted and displayed correctly.

+++ firmware/base_esp32/packet.h
@@ -6,14 +6,15 @@
 #define GPS_FLAG_FIX_MASK 0x80  // Bit 7: Fix válido
 #define GPS_FLAG_SATS_MASK 0x7F // Bits 0..6: Satélites visibles
 
+// Estructura binaria compacta (23 bytes totales)
 struct __attribute__((packed)) MinimalCollarPacket {
-  uint16_t collar_id;  // 2 bytes: Identificador de nodo (0x9E2B)
-  uint16_t seq;        // 2 bytes: Secuencia incremental
+  uint64_t collar_id;  // 8 bytes: Identificador de 16 caracteres hex
+  uint16_t seq;        // 2 bytes: Contador secuencial
   int32_t lat_scaled;  // 4 bytes: Latitud * 10,000,000
   int32_t lon_scaled;  // 4 bytes: Longitud * 10,000,000
   int16_t alt_m;       // 2 bytes: Altitud en metros
   uint16_t battery_mv; // 2 bytes: Batería en milivoltios
-  uint8_t gps_flags;   // 1 byte:  Bit 7: Fix | Bits 0..6: Satélites
+  uint8_t gps_flags;   // 1 byte:  Bit 7 Fix | Bits 0..6 Satélites
 };
 
 #endif // PACKET_H
